refactor: changed logic in user_handler
- Added costs on buttons in menu_recall

+++ bot/user_handler.py
@@ -3,14 +3,8 @@ from pprint import pprint
 import random
 random.seed(42)
 
-from aiogram import Router, F, types, Bot, exceptions
-# from aiogram.enums import ParseMode
-# from aiogram.types import Message, message, pre_checkout_query, user
-# from aiogram.fsm.state import State, StatesGroup
+from aiogram import Router, types, Bot, exceptions
 from aiogram.filters import Command
-# from aiogram.utils.keyboard import InlineKeyboardBuilder
-# from aiogram.types import InlineKeyboardMarkup
-# from aiogram.exceptions import TelegramBadRequest
 
 from data.presets import FTEXT
 from core.rogue_interface import RogueInterface, process_action, NoAction
@@ -26,7 +20,8 @@ class UserController:
 
         self._register_handlers()
 
-        self.lock_users = set()
+        self.lock_users: set[int] = set()
+        self.lock_messages: set[int] = set()
 
     def _register_handlers(self) -> None:
         """
@@ -70,7 +65,8 @@ class UserController:
                     )
                 return
             except (exceptions.TelegramBadRequest, exceptions.TelegramForbiddenError) as e:
-                asyncio.create_task(self._delete_bot_message(message_id=ui_message, chat_id=user_id, delay=3))
+            # except Exception as e:
+                asyncio.create_task(self._delete_bot_message(message_id=ui_message, chat_id=user_id, delay=2))
 
         if map_photo is not None:
             new_ui_message = await self.bot.send_photo(
@@ -88,6 +84,7 @@ class UserController:
         await self.interface.save_ui_message_id(user_id, new_ui_message.message_id)
 
     async def _delete_object_message(self, *, message: types.Message, delay: int = 0, warning_message: types.Message | None = None) -> None:
+        self.lock_messages.add(message.message_id)
         await asyncio.sleep(delay)
         try:
             await message.delete()
@@ -95,13 +92,18 @@ class UserController:
                 await warning_message.delete()
         except:
             pass
+        finally:
+            self.lock_messages.remove(message.message_id)
 
     async def _delete_bot_message(self, *, message_id: int, delay: int = 0, chat_id: int) -> None:
+        self.lock_messages.add(message_id)
         await asyncio.sleep(delay)
         try:
             await self.bot.delete_message(message_id=message_id, chat_id=chat_id)
         except:
             pass
+        finally:
+            self.lock_messages.remove(message_id)
 
     # === Commands ===
 
@@ -109,25 +111,63 @@ class UserController:
         if message.from_user is None: return
         user_id = message.from_user.id
 
-        contract = await self.interface.cmd_start(user_id)
+        if user_id in self.lock_users:
+            await self._delete_object_message(message=message)
+            return
+        self.lock_users.add(user_id)
+        try: # To exactly unlock the user
+            contract = await self.interface.cmd_start(user_id)
 
-        keyboard = get_keyboard(contract.buttons)
-        cmd_start_text = random.choice(FTEXT["cmd_start"])
+            keyboard = get_keyboard(contract.buttons)
+            cmd_start_text = random.choice(FTEXT["cmd_start"])
 
-        # await message.answer(cmd_start_text, reply_markup=keyboard)
-        asyncio.create_task(self._delete_object_message(message=message, delay=3))
-        await self._update_bot_message(user_id, cmd_start_text, keyboard, contract.map_photo)
+            asyncio.create_task(self._delete_object_message(message=message, delay=3))
+            try:
+                ui_message = await self.interface.get_ui_message_id(user_id)
+                await self._delete_bot_message(message_id=ui_message, chat_id=user_id)
+
+            except ValueError:
+                pass
+
+            ui_message = await self.bot.send_message(
+                chat_id=user_id,
+                text=cmd_start_text,
+                # reply_markup=keyboard
+            )
+            await asyncio.sleep(2)
+
+            await self.bot.edit_message_text(
+                        chat_id=user_id,
+                        message_id=ui_message.message_id,
+                        text=cmd_start_text,
+                        reply_markup=keyboard
+                    )
+
+            await self.interface.save_ui_message_id(user_id, ui_message.message_id)
+        finally:
+            self.lock_users.remove(user_id)
+            self.lock_messages.remove(message.message_id)
 
     async def cmd_expanse(self, message: types.Message) -> None:
         if message.from_user is None: return
         user_id = message.from_user.id
-        contract = await self.interface.goto_menu_expanse(user_id)
 
-        keyboard = get_keyboard(contract.buttons)
-        menu_text = contract.text
+        if user_id in self.lock_users:
+            await self._delete_object_message(message=message)
+            return
+        self.lock_users.add(user_id)
 
-        await self._delete_object_message(message=message)
-        await self._update_bot_message(user_id, menu_text, keyboard, contract.map_photo)
+        try: # To exactly unlock the user
+            contract = await self.interface.goto_menu_expanse(user_id)
+
+            keyboard = get_keyboard(contract.buttons)
+            menu_text = contract.text
+
+            await self._delete_object_message(message=message)
+            await self._update_bot_message(user_id, menu_text, keyboard, contract.map_photo)
+        finally:
+            self.lock_users.remove(user_id)
+            self.lock_messages.remove(message.message_id)
 
     # async def cmd_help(self, message: types.Message) -> None:
     #
@@ -151,9 +191,10 @@ class UserController:
 
     async def callback_handler(self, callback: types.CallbackQuery) -> None:
         user_id = callback.from_user.id
-        if user_id in self.lock_users:
+        message_id = callback.message.message_id # type: ignore
+        if user_id in self.lock_users or message_id in self.lock_messages:
+            await callback.answer()
             return
-
         self.lock_users.add(user_id)
 
         try:
@@ -164,7 +205,6 @@ class UserController:
             keyboard = get_keyboard(contract.buttons)
 
             pprint(contract)
-            # await callback.message.answer(contract.text, reply_markup=keyboard) #type: ignore
             await self._update_bot_message(user_id, contract.text, keyboard, contract.map_photo)
 
         except NoAction:
